rename simple browser service
